refactor(eve): align tool source boundaries
