+++ packages/eve/extension-contracts/entrypoints/dynamicTool.ts
@@ -4,4 +4,4 @@ export {
   type DynamicToolResult,
   type DynamicToolSet,
   defineDynamic,
-} from "../../src/tools/index.ts";
+} from "../../src/public/tools/index.ts";
