+++ packages/eve/src/internal/nitro/routes/agent-info/build-agent-info-response.test.ts
@@ -4,7 +4,7 @@ import { compileFromMemory } from "#compiler/compile-from-memory.js";
 import { buildAgentInfoResponse } from "#internal/nitro/routes/agent-info/build-agent-info-response.js";
 import { defineInstrumentation } from "#public/instrumentation/index.js";
 import { experimental_workflow } from "#public/definitions/tool.js";
-import { webSearch } from "#tools/builtins/web-search.js";
+import { webSearch } from "#public/tools/web-search.js";
 
 describe("buildAgentInfoResponse", () => {
   it("projects v3 exclusively from the effective compiled graph", async () => {
