+++ packages/eve/src/public/tools/bash.ts
