+++ packages/eve/src/public/tools/read-file.ts
