+++ packages/eve/src/tools/agent.ts
