+++ scripts/extension-contracts/configuration.mjs
@@ -22,7 +22,11 @@ export const PUBLIC_SURFACES = [
   { path: "src/public/schedules/index.ts", capabilities: ["schedule"] },
   { path: "src/public/index.ts", capabilities: ["subagent"] },
   {
-    paths: ["src/tools/index.ts", "src/tools/workflow.ts", "src/tools/builtins/web-search.ts"],
+    paths: [
+      "src/public/tools/index.ts",
+      "src/public/tools/workflow.ts",
+      "src/public/tools/web-search.ts",
+    ],
     capabilities: ["tool", "dynamicTool"],
   },
   { path: "src/public/connections/index.ts", capabilities: ["connection"] },
